refactor: investment 관련 API를 investmentApi로 분리 및 사용처 수정
